+++ src/VendingMachine/Business/Exceptions/WalletDoesNotHaveCoinsForParValueException.cs
@@ -5,9 +5,9 @@ namespace VendingMachine.Business.Exceptions
 {
     public sealed class WalletDoesNotHaveCoinsForParValueException : Exception
     {
-        public readonly int ParValue;
+        public readonly decimal ParValue;
 
-        public WalletDoesNotHaveCoinsForParValueException(int parValue) 
+        public WalletDoesNotHaveCoinsForParValueException(decimal parValue) 
             : base(string.Format("Wallet does not have coins with par value {0}", parValue))
         {
             Contract.Requires(parValue > 0);
